Add files via upload
COAL Project.
We used the Arduino UNO as a microprocessor, sensor for the traffic motion is IR tracking sensor. That sense the motion of cars on the road, We used RGB LED for the traffic lights, and a simple blue LED for the pedestrian Light, button has been used in the circuit, so basically our circuit ill not work until the button is pressed that means that this traffic system can be turned on and off when ever wanted.